+++ src/main/java/net/javaguides/springboot/springsecurity/controller/PackageController.java
@@ -1,6 +1,7 @@
 package net.javaguides.springboot.springsecurity.controller;
 
 import lombok.RequiredArgsConstructor;
+import net.javaguides.springboot.springsecurity.ForbiddenException;
 import net.javaguides.springboot.springsecurity.model.PackageDto;
 import net.javaguides.springboot.springsecurity.model.dto.ChangePackageRequest;
 import net.javaguides.springboot.springsecurity.service.PathFileService;
@@ -13,21 +14,27 @@ public class PackageController {
 
     private final PathFileService pathFileService;
 
-
     @GetMapping("{name}")
-    public PackageDto getPath(@PathVariable("name") String name) {
-
-
+    public PackageDto getPath(@PathVariable("name") String name, @RequestHeader("token") String token) {
+        if(token.isEmpty()){
+            throw new ForbiddenException("Token is null");
+        }
         return pathFileService.getPath(name);
     }
 
     @DeleteMapping("{name}")
-    public void deletePackage(@PathVariable("name") String name) {
+    public void deletePackage(@PathVariable("name") String name, @RequestHeader("token") String token) {
+        if(token.isEmpty()){
+            throw new ForbiddenException("Token is null");
+        }
         pathFileService.deletePackage(name);
     }
 
     @PutMapping
-    public void changePackage(@RequestBody ChangePackageRequest changePackageRequest) {
+    public void changePackage(@RequestBody ChangePackageRequest changePackageRequest, @RequestHeader("token") String token) {
+        if(token.isEmpty()){
+            throw new ForbiddenException("Token is null");
+        }
         pathFileService.changePackage(changePackageRequest);
     }
 }
